--- srcs/init_utils/parse_cubfile_utils/find_next_line.c
@@ -1,14 +1,0 @@
-#include "cub3d.h"
-
-char	*find_next_line(const char *contents)
-{
-	char	*line;
-
-	line = ft_strchr(contents, '\n');
-	if (line == NULL)
-		return (NULL);
-	line++;
-	if (*line == '\0')
-		return (NULL);
-	return (line);
-}
