+++ srcs/init_utils/parse_cubfile_utils/parse_map_utils/get_player_info.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   init_player.c                                      :+:      :+:    :+:   */
+/*   get_player_info.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 01:17:48 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/11/04 23:32:53 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/11/05 14:47:44 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
