+++ srcs/init_utils/parse_cubfile_utils/parse_map_utils/set_direction.c
